print a number by digits

+++ 0x06-pointers_arrays_strings/101-print_number.c
@@ -6,13 +6,13 @@
  */
 void print_number(int n)
 {
-	int number = n;
+	unsigned int number = (unsigned int)n;
 	unsigned int tens = 10;
 
 	if (n < 0)
 	{
 		_putchar('-');
-		number = (int)(((unsigned int)number) * ((unsigned int)-1));
+		number = (unsigned int)((int)(((unsigned int)number) * ((unsigned int)-1)));
 	}
 	while ((number / tens) >= 10)
 		tens *= 10;
